ensure type export
